[iOS] Migrate away from TestChromeBrowserState::CreateBookmarkModel()
This patch migrates tests to use builder APIs instead of the deprecated
TestChromeBrowserState::CreateBookmarkModel().

Following the approach in chrome/, BookmarkModelFactory is extended
with a function that exposes the default factory, for tests that
are insterested in instantiating BookmarkModel.

The resulting behavior is identical. Note that the new factory function
integrates with ManagedBookmarkServiceFactory, as opposed to
TestChromeBrowserState which always used null for
ManagedBookmarkService. That said, this makes no difference in practice
because the factory uses null by default in tests.

Bug: 1106699
Change-Id: Ifbffb3e6921d66868b419614851727a2d422eee5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3384557
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#960069}
NOKEYCHECK=True
GitOrigin-RevId: adb2668e63ccd5d4117b2dc671c076f2ff9a5e91