[SH 2.0 iOS] Add additional EG tests
This CL add a test for the "Remove Highlights" flow, and one verifying
that closing the menu is a no-op.

A new helper method is added to ChromeEarlGrey, for verifying that a
given element is *not* present in a WebState. This is needed to verify
that the <mark> elements have been removed.

Bug: 1286018
Change-Id: I152e0720315a4247400be9b7fb6d969bbfadf669
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3383595
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#960465}
NOKEYCHECK=True
GitOrigin-RevId: 5648f46934d8aa3638b9308929723b795f60eacd