[ios] Add initWithConfiguration: to Most Visited and Return to Recent Tab
Adding initWithConfiguration: as an initializer for Content Suggestions
views will set the foundation for the Content Suggestions migration
to UIViewController. Currently, configureCell: configures each element's
view. With initWithConfiguration:, the necessary configurations will
be passed to the view for it to self-configure.

This requires the views to have references to the configuration objects,
which creates a dependency cycle between //ntp_tile_views and
//content_suggestions/cells, requiring them to be merged. This merge
also necessitates //content_suggestions/cells:cells_ui merge in
as well.

Bug: 1285378
Change-Id: I33bec49a096b800990dee0cebfe0ed115449ba7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3382314
Reviewed-by: Adam Trudeau-Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#960467}
NOKEYCHECK=True
GitOrigin-RevId: 7b77c27f789ed9dbe15e0525edc904648b4016ae