[iOS] Fix CPE transient authentication error
Occasionally, when |-prepareCredentialListForServiceIdentifiers:| is
called, if we attempt to authenticate the user using the iOS
authentication framework, the authentication immediately fails. It seems
that iOS thinks that the extension is not foregrounded and so can't
authenticate.

This CL (hopefully) fixes the issue by waiting slightly to authenticate
until the view controller appears, ensuring the extension is
foregrounded.

Bug: 1273460
Change-Id: I1528275fd062de520b53bfc043efd2cc7b485ee9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3399510
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#960656}
NOKEYCHECK=True
GitOrigin-RevId: 9f4169891fd2155781ba143446dadcccf2d8925c