Remove switches::kDisableSyncTypes
The only use of the flag was to disable a data type that's been
removed, see linked bug. This leads to simplifications in
ChromeSyncClient and IOSChromeSyncClient, where individual types
can no longer be disabled. Unfortunately, the disabling must still
be supported by the CreateCommonDataTypeControllers() API, because
iOS web view relies on this to allowlist a handful of types.
https://source.chromium.org/chromium/chromium/src/+/main:ios/web_view/internal/sync/web_view_sync_client.mm;l=44;drc=1c54d1a6d5c714cc07c19c255b0832c51e392330

I'll add that
a) A few ChromeOS data types in ChromeSyncClient were not respecting
the disabled_types parameter, which could have been a source for
bugs. So removing the parameter is a win. It also makes the
CreateDataTypeControllers() method marginally easier to read.
b) It is weird that we have some concept of disabled_types
in a CreateCommonDataTypeControllers() API. Maybe the
types that can be disabled should be moved out of this method.

Fixed: 1288789
Change-Id: If05345a65c63c34999e9d4be8037660aeb866b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3399838
Auto-Submit: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#960960}
NOKEYCHECK=True
GitOrigin-RevId: f420f88edac8c82ffbd7f6029fdb241294b1e0b0