[SH iOS] Add JS events to support taps on highlights
Currently, we fire an event when the user taps anywhere on the page.
With this patch, a click event is attached to each highlight in the
page. When tapped, a new event (onClickWithSender) is piped through,
including the coordinates of the sending highlight, up to the ui layer.

This causes two behavior changes in the UI when the
kIOSSharedHighlightingV2 flag is on:

* The menu only pops up when the user taps a highlight, not other parts
  of the page.
* On tablets, the menu is anchored to the coordinates of the text the
  user tapped on.

There is no behavior change when the flag is off.

As a transition state, we pipe *both* the page-level taps and the
highlight-level taps, and the mediator responds to one or the other
depending on a flag value. When this new behavior is enabled by default,
we can remove the version triggered on page tap (i.e., the version that
does not include withSender).

Fixed: 1286019
Change-Id: I8fb0f9154fa54c722b47a3d983fc16a38771c764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3404759
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#962009}
NOKEYCHECK=True
GitOrigin-RevId: c2af76f7217094199a6ff31a3d68a9a982f9ce5d