[iOS] Reland -- Factor BVC's delegate installation into a mediator
Factor BVC's web state delegate installation into a mediator, part one
of many.

Currently the BVC is responsible for setting the delegates of several
tab helpers with UI dependencies. This requires the BVC to observe the
web state list to update these helpers when new tabs are added, which
is outside of a view controller's responsibilities.

The BVC also implementing these delegate protocols is also a problem,
but not one addressed by this series of CLs.

This CL factors this job into a mediator (TabLifecycleMediator) which
the BrowserCoordinator owns. It's responsibility is to  handle injecting
UI dependencies into new web states (and removing them from departing
ones). To do this, the mediator will be provided with a slew of
dependencies (factories and other objects) as well as a catchall
delegate which conforms to all of the tab helper delegate protocols
that the BVC implements. The mediator uses the
WebStateDependencyInstaller abstraction to handle observation and
triggering of install/uninstall events; the new mediator simply
implements the necessary methods.

Most of this CL is setting up the infrastructure for migrating the
tab helper installations. Only one tab helper is moved into the mediator
(SnapshotTabHelper). Two other tab helpers which had no reason to be
initialized in the BVC (OfflinePageTabHelper and the translate tab
helper) have their init moved to AttachTabHelpers().

Future CLs will move more of the tab helpers initialized in the BVC and
BrowserCoordinator into TabLifecycleMediator. After that, further
refactoring can consolidate the model and UI-layer code for each feature,
removing the delegate responsibilities from the BVC.

To facilitate this, this CL adds a category on the BVC which includes
its conformance to the various tab helper delegate protocols. This
category is only used by the BVC and BrowserCoordinator.

Bug: 1290819
Change-Id: I4cc04fd6003d9bf89049b0646cc2eb1be6d9c66d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3326281
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#964547}
NOKEYCHECK=True
GitOrigin-RevId: 0fc927c722b9356f7743b383767d442879dad205