[ios] Avoid creating a TestChromeBrowserState in FakeSceneState
FakeSceneState is an Objective-C object meaning that if it
creates a TestChromeBrowserState, the C++ object may outlive
the objects it depends on (such as web::WebTaskEnvironment).

Instead change FakeSceneState to take the ChromeBrowserState
to use in its initializer (and fix the incognito browser to
use an incognito ChromeBrowserState instead of a second
TestChromeBrowserState).

Bug: 1290336
Change-Id: I579404a69b696d5f290d9b2c69331679469c00e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3421425
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#964962}
NOKEYCHECK=True
GitOrigin-RevId: cee972dbeee2f9ca51bf7333b4912e7cc5ad95af