[ios] Avoid creating a BrowserState in TestBrowser
In production, Browser is (indirectly via BrowserList) owned
by ChromeBrowserState. Having TestBrowser create and own the
ChromeBrowserState means that the lifetime of the objects are
different between production and tests.

Having different lifetime relationship means that tests may
not reflect what is happening in production and may hide some
behaviour that are incorrect behind the different lifetimes.

Change all tests to create explicit the TestChromeBrowserState
and remove the default constructor of TestBrowser.

Bug: 1290337
Change-Id: I0a3a606fb41ef3e3037dfe32aefbaa012a53a9f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3420297
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#964963}
NOKEYCHECK=True
GitOrigin-RevId: e5c3e764c2530c143717f4056af4ac24036d7120