[iOS] Move more tab helper setup to lifecycle mediator.
This CL moves setup for three classes out of the BVC and into the
TabLifecycleMediator, Specifically it moves the three objects that
have dependencies on BrowserViewController properties, so in order to
facilitate this refactor, those properties are exposed in the +Delegates
BVC category, and then passed into the TabLifcycleMediator as
dependencies.

Dealloc timing issues in DownloadManagerCoordinator are also fixed (since
the coordinator may dealloc after -stop is called).

As noted in the various TODOs in this CL, further refactoring needs to
be done to completely extract these features from the BVC.

Bug: 1290819
Change-Id: I4396cdb70bc41c5d6eb874cde63709f0dec31188
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3420566
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#965273}
NOKEYCHECK=True
GitOrigin-RevId: dfbfa1dbb6061f68bc18d538f67a2a8e261b306f