[ios] Stop access ApplicationContext before creation
The ApplicationContext is created when the application enter the
InitStageBrowserObjectsForBackgroundHandlers (in MainController).
Before that point many Chrome objects are not available and cannot
be used.

When the OS decides to garbage collect "scenes" while the app is
not running, it sends -application:didDiscardSceneSessions: the
next time the application is started. This message is sent
before -sceneWillConnect: which means that the app may still not
be in InitStageBrowserObjectsForBackgroundHandlers (or above).

The ChromeIdentityService may not yet be accessible by that point
as it depends on the ApplicationContext being created. Since the
purpose of -application:didDiscardSceneSessions: is to allow the
application to clean up its state tied to those sessions, it is
fine if the deletion is delayed until the application has finished
correctly initialized.

Since observing the AppState's InitStage changing requires to be
an AppStateObserver, put the observation in MainApplicationDelegate
(observing AppState in AppState is an anti-pattern according to
a comment in AppState).

Fixed: 1292641
Change-Id: I75a378be3782d2da6c171dcfc926be3578382620
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3431645
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966660}
NOKEYCHECK=True
GitOrigin-RevId: 18154991c809c7e482d37ba5df0923a00941ee3a