[iOS] Move ManageAccountsDelegate out of BVC
This CL factors ManageAccountsDelegate out of the BVC and into a
browser agent. Because the BVC's implementation of the protocol
didn't depend on any BVC internals, the refactor is straightforward.

The new AccountConsistencyBrowserAgent now handles both adding and
removing web states to the AccountConsistencyService, and acts as the
delegate used by the service for handling account events.

Because the browser agent is a C++ class, an internal Objective-C
bridge class is used to forward protocol calls into the browser agent.

A follow-up CL should update the ManageAccountsDelegate protocol to
be a C++ abstract class instead.

Fixed: 1272476
Change-Id: I3ab16ac57f1c617bb066ad7a570e641a4799de7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3431428
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966711}
NOKEYCHECK=True
GitOrigin-RevId: f793d736a1e481d6639e9af4dee31dc3f314e9d0