[ReadingList] Subclass KeyedService in ReadingListModel
ReadingListModel is almost a pure interface which means that
it is easy to create a fake object for tests. However since
it does not sub-class KeyedService, the factory cannot cast
the KeyedService* to ReadingListModel. This means that it is
not possible to set a testing factory that returns a fake
object that does not sub-class ReadingListModelImpl.

Instead, change ReadingListModel to subclass KeyedService.
This way, the factory can be changed to not depends on the
created object to be a ReadingListModelImpl, allowing to
instead install e.g. a FakeReadingListModel.

Bug: none
Change-Id: I266a70db5f5fa93f1c45e439c5ac080c37e8a6d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3440128
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967246}
NOKEYCHECK=True
GitOrigin-RevId: 48f7b191fb6aee2bb94d3da3d19a42eaf978f17c