[iOS] Change font for popover label
This CL changes the font of the popover label to match with enterprise.

The blur effect is already applied to non-enterprise popover labels, so
it would not be changed in this CL. The original bug mentioned blur
effects due to a display discrepancy between Figma and Apple.

Fixed: 1284700
Change-Id: I6821151c47a88164a2096acebb085909cb7ba4a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3432187
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967288}
NOKEYCHECK=True
GitOrigin-RevId: c882cc3c304007c37ad6d5767066091e3342d265