[iOS] Disable enterprise launch screen in test
After https://crrev.com/c/3217455, iphone_simulator and ipad_simulator
started to failed.

The tests were about testing if the device properly fetch policies and
display that the device is managed.  The enterprise launch screen is
displayed at first run when the device is managed and policies are
fetched.

To CL, add a new check similar to the other first run screens about if
the first run should be displayed or not as the enterprise launch screen
is part of first run.

Bug: 1294334
Change-Id: I01e31f8179243a849cc4a12937faf020dbf46a2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3443345
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@google.com>
Cr-Commit-Position: refs/heads/main@{#967807}
NOKEYCHECK=True
GitOrigin-RevId: dce4871b12c5c3133cd94f854e12090d2f79dcc6