[iOS] Manually show new permissions banner when old banner is dismissed
Currently, if the user enables camera permission and dismisses the
resulting banner, when they enable a new permission, the infobar banner
would not show again.

This is because of the way the overlay request cancel handler handles
replacement. A new request for the replacement's banner is only inserted
if an infobar is replaced while its banner request is in the queue.

This CL makes an exception for permissions - a new request for the
updated infobar is manually added when its banner request is NOT in the
queue.

Fixed: 1295833
Change-Id: If6c02489ac31d8dc6f440855ce74d65f2f4ca002
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3449587
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#969642}
NOKEYCHECK=True
GitOrigin-RevId: ae8546db2a4d3d4a5fbe4f40135d4c19c610055d