[ios] Reference XCTest and XCUITest in generated xcscheme
To allow running XCTest in Xcode, the scheme for the test host
needs to list all of the XCTest modules targets that uses it as
and host. This is also needed for XCUITests to run them via the
host and not via the test module target.

Update convert_gn_xcodeproj.py script to include all the test
module targets in their corresponding test host scheme.

Bug: none
Change-Id: I10154cf8840b3f31b8399dc6cf0dd2f7e63c14df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3461356
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970653}
NOKEYCHECK=True
GitOrigin-RevId: 3973d78849b862b2ff7d7f7ed73bc91b01997ba7