[iOS] Fix omnibox copying color in incognito mode
The root cause of the bug is that the interaction is attached to
_locationBarSteadyView, which is transparent. iOS automatically adds
a light background in light mode.

The fix is to manually change the target view for the interaction so
that the view with the correct background color (the location bar
view controller's super view) is the view used for the preview.

This still has a tiny bug where the white background can be seen around
the rounded corners of the location bar container view, but in my
opinion, this isn't a huge deal.

Bug: 1292993
Change-Id: I103396dc2e75fe1929581d6cd600444d3a660c0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3457367
Auto-Submit: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971182}
NOKEYCHECK=True
GitOrigin-RevId: 1fcf1109d922c89042ea9fad22206c9a4f719c4c