[ios] Move header out of ContentSuggestions
When kContentSuggestionsHeaderMigration is enabled, the header
containing the fake omnibox, logo, and avatar will be pulled out of the
ContentSuggestions and instead managed by the NewTabPage logic. In the
initial state, the header will be placed in the Discover Feed ScrollView
above the ContentSuggestions.

Bug: 1285378
Change-Id: Ia0e982101e7249c3ac0bdaaee67bb144e8783f77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3373711
Reviewed-by: Adam Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971300}
NOKEYCHECK=True
GitOrigin-RevId: 9199aef1b18b70dc94f25fccae14b6f164eef68d