[iOS][media-permissions] Adds permissions badge in location bar
This CL adds a permissions badge in the location bar. To do so, two more
|BadgeType|s are added: kBadgeTypePermissionsCamera and
kBadgeTypePermissionsMicrophone.

Note that the icon and acceptance of the location bar depends on the
state of permissions retrieved from the web state, therefore when badge
items are created in the mediator's |self.badges|, the BadgeType of
InfobarTypePermissions is not retrieved from BadgeTypeForInfobarType,
but set separately based on |self.permissionsBadgeType|.

The mediator also starts observing the active web state, and will
explicitly trigger a badge re-render when any permission state changes.

Bug: 1286892
Change-Id: I41f903f09d3a58c504815f0e263ac49fb11dd17a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3457384
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: edchin <edchin@google.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971415}
NOKEYCHECK=True
GitOrigin-RevId: fd01512ede5ebd6ec8673cc561caa5b8343e7190