[ios] Update parent item when a Most Visited favicon is retrieved.
When kSingleCellContentSuggestions is enabled, the parent item is the
one used by the CollectionView to configure the entire
ContentSuggestions. Thus, a new API is needed for
ContentSuggestionsFaviconMediator to pass the parent item in order to
update it when a favicon is fetched.

Bug: 1296224
Change-Id: I69d7ea25b02afb7476d8b4e0f3460239092bf212
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3453871
Reviewed-by: Adam Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971421}
NOKEYCHECK=True
GitOrigin-RevId: 2b4660ecb290727ec09f50a7c7e78cb189d09614