[iOS] Remove new overflow menu from tab grid snapshot
There are 2 issues solved in this CL.

First, the overflow menu was being presented by the BVC (which ends up
being the tab grid as the parent that defines presentation context).
However, |presentedViewController| walks up the hierarchy, even over
a defined presentation context, so the overlay container view controller
thinks it is presenting something. This was fixed by requiring the
overlay container view controller's |presentedViewController|'s
|presentingViewController| to be the overlay VC itself.

The second issue is that iOS automatically dims the tint color of views
behind a modally presented view. Thus, the snapshot had gray instead
of blue there. To fix this, the snapshot generator now changes the
view's |tintAdjustmentMode| before snapshotting to remove the dimming.
And unfortunately, there's no way to tell if a view's
|tintAdjustmentMode| was manually changed or is being inherited from
a superview, so this also adds a new delegate method to
|SnapshotGeneratorDelegate| so the |tintAdjustmentMode| can be set
back correctly.

Bug: 1292459
Change-Id: I095fea16b8e9af01b1e32dd6ff7274d7cd6e3f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3428825
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: edchin <edchin@google.com>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#971997}
NOKEYCHECK=True
GitOrigin-RevId: 6cea9cdb5271fe546ff3b6fade16e49e0704ccfc