[SH 2.0 iOS] Reorder menu items
The items in the menu need to be reordered to conform to the Human
Interface Guidelines. This new order has been approved by UX.

The diff doesn't make this super clear, but there are no changes to the
contents of the code--this is a strictly cut-and-paste job to rearrange
the lines.

Bug: 1298085
Change-Id: I76ad1797093781f9ae43c36929a6d0b8b22f6fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469179
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Auto-Submit: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/main@{#972126}
NOKEYCHECK=True
GitOrigin-RevId: c8d633e4211c2f73af51d782d8d2e523a7d7d4b5