[WebChannels] Return FollowSiteInfo with none rss links
Previously when there's no rss links fetched, we return a nil as
siteInfo, but the FollowSiteInfo object is still useful when other
properties are not nil. This change returns the site info with a nil
RSSlinks when no rss links is fetched.

Bug: 1264872
Change-Id: Ieab1d60f0da4ecce7443731af6a5dce35d1681e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3463023
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972214}
NOKEYCHECK=True
GitOrigin-RevId: cd00a22825c6724f068ae7f47c178599b242ef98