[ios] Potential fix for crash in net::URLFetcherFileWriter::Write
The crash appears to happen when net::URLFetcherFileWriter::Write
is called after net::URLFetcherFileWriter::Finish which may happen
if a DownloadSessionTaskImpl is reused for a new download.

Clear the session_task_ member field before calling Finish, and
check the task we get the event from is the same as session_task_
to avoid getting in that state.

Also use base::BindRepeating() with member functions instead of
blocks to avoid capturing `this` implicitly. This also allow to
check whether the DownloadSessionTaskImpl has been destroyed and
to cancel the task in that case.

Bug: 929112
Change-Id: I060901d336a073b99f1374ff399f0bb0f4dd1427
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3465194
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972881}
NOKEYCHECK=True
GitOrigin-RevId: 3af0eed7801f0e65707a312bfe31bdbcbc69d47b