[iOS] Use SwiftUI popup
Creates a hosting controller for the swiftui popup view, and uses it
when the feature flag is on.
Moves the preview project and sources to popup/ and splits it to
shared and non-shared code.
Adds a README to this effect.

PAIR=rkgibson, qpubert

Fixes: 1298940
Change-Id: I87704c280b555e43780a7c89b7a43eaa321120c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468773
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973109}
NOKEYCHECK=True
GitOrigin-RevId: 42061f2e78baea763821c6d08d631f4f1921b6a1