[ios] Fix AnimatedScopedFullscreenDisabler
FullscreenController is a base::SupportsUserData::Data attached to a
Browser, like the owner of AnimatedScopedFullscreenDisabler. There is
no guarantee in which order the base::SupportsUserData::Data are to
be destroyed.

Observe FullscreenController in AnimatedScopedFullscreenDisabler (as
it can notify that it will be destroyed) and set `controller_` to
null when this happen (this likely mean that the application is to
be destroyed).

Also use base::WeakPtr<AnimatedScopedFullscreenDisabler> to prevent the
blocks passed to -[UIView animateWithDuration:animation:completion:]
from capturing pointers to C++ objects (can lead to use-after-free).

Bug: 1241310
Change-Id: I6cf37599af0b9861d10c6414cf5a55e9f4ad8adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3478034
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973661}
NOKEYCHECK=True
GitOrigin-RevId: 18b49fbeccf44cbef9bd383e32204667bc46b9e9