Add suggested actions section to the normal tabs page of Tabgrid search
1- Created "SuggestedActionsGridCell" as a new cell type to be used in
the gridViewController and will be added under a separate new section
"SuggestedActionsSection" in the end of the grid view.
2- Created "SuggestedActionsViewController" as ChromeTableViewController
which will contains the suggested actions items.
3- GridViewController will create and own an instance of
SuggestedActionsViewController and will implement its delegate. It will
also provide its tableView to the SuggestedActionsGridCell upon its
creation.
4- SuggestedActionsGridCell contains a tableView of all the suggested
actions. That table view is rendered as a subview of the cell and its
boundaries matches the cell boundaries.
5- "SuggestedActionsDelegate" is added to handle the actions, is called
by The gridViewController, and implemented by the TabGridViewController.
6- Updated the TabGridViewControllerDelegate with two methods to handle
suggested actions, and added their implementation to the
TabGridCoordinator.
7- Added a gridCommand (implemented by TabGridMediator) to fetch the
history search results count, to be used inside the history cell in the
suggestedActionsViewController.
8- Added some unittests for the newly created view controller. Tests for
the UI will done on a following CL as an EG test.
9- The whole change is guarded by a new feature flag which will be
enabled by default "TabsSearchRegularResultsSuggestedActions" and
requires that TabsSearch flag to be enabled as well.

screenshots:(Another change will be needed to add a header to the
results)
https://screenshot.googleplex.com/3iEswCBGwcxZvdi.png
https://screenshot.googleplex.com/3n4MK9WVwFb5Wsv.png

Bug: 1290663
Change-Id: I067fb808ea25bbfedf50572b16c7f37f687b6895
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3476714
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974015}
NOKEYCHECK=True
GitOrigin-RevId: 9541008f8153209014e0be29583d66f80d0a1796