[iOS] Change isFirstRun value before enterprise stage
To block landscape on enterprise launch screen, the `portraitOnly`
function, in app state, checks if the current stage is InitStageFirstRun
or InitStageEnterprise and if this is the first run. But the first run
value is set after enterprise stage.

This CL:
* Set first run value for InitStageEnterprise
* Correct tests accordingly

Fixed: 1297299
Change-Id: I0258ea3905ffb750654ddf87e0035a2a30d26791
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3487909
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@google.com>
Cr-Commit-Position: refs/heads/main@{#975052}
NOKEYCHECK=True
GitOrigin-RevId: ca948aac35cc0a943c2631dfb1542b8154c86106