Fix snapshots not updating correctly in search mode
There were a problem were the cells are not updated when the collection
view is updated (reloaddata was basically no-op). Right now reloaddata
only works correctly and reload all cells if the collection view was
empty. Otherwise it has to be called 1 cycle later.

I tried to use performBatchUpdates to grantee that reload data is called
in the correct time. That worked fine and reloaded the snapshots
correctly. However the app initialization depend on that populateItems
method is called first thing and with performBatchUpdates this can't be
granteed if another performBatchUpdates was called in the same thread
cycle.

To fix the problem for now, the mediator will populateItems in a
dispatch_async to ensure that it runs on the UI thread and also wait for
the current thread cycle to end. i created a crbug/1300733 to followup
on the investigation of how to fix populateItem to work correctly in all
cases.

Fixed: 1299417
Change-Id: I337a90e27ac2ac51fc430a9ae1b5cc13ab90d732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3486202
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#975131}
NOKEYCHECK=True
GitOrigin-RevId: e6dadc1554fe3b20b645b43ec8b60da3fd19fd26