[ios] Create tab helper in AttachTabHelpers
Create NetExportTabHelper and RepostFormTabHelper, PassKitTabHelper
and DownloadManagerTabHelper in AttachTabHelpers to ensure they are
created as soon as a WebState is inserted in a WebStateList.

Add explicit method SetDelegate() to those classes instead of passing
the delegate in the constructor. Call SetDelegate() from the location
where the tab helper were previously created.

Fix a possible Use-After-Free in RepostFormTabHelper.

Fix the unit tests.

Bug: 1300151
Change-Id: I501df87b269c611526b4d4d20c6f2df11c0cbc65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3484287
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#975722}
NOKEYCHECK=True
GitOrigin-RevId: 62cb91ccefcbc775084b8ca4207ca07f75427968