[iOS] Use PopupModel as AutocompleteResultConsumer, use its delegate.
Makes PopupModel a result consumer. Moves the creation and the ownership
of the model to the Coordinator. The Mediator now knows no difference
between the swiftui and uikit consumers.

Adjust the consumer protocol to remove GURLs from the swiftui version.

Add a delegate to PopupModel which is a ACResultConsumerDelegate. Use
that for deletion instead of a closure.

Bug: 1298940
Change-Id: I946636a1fb07124867310773b8bfbd5c102083dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3481338
Auto-Submit: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#975922}
NOKEYCHECK=True
GitOrigin-RevId: 4c9b53c6aa99c57dbeddd20000f1493be0d35057