[iOS] Close password suggestion when app backgrounds.
Currently, the password suggestion sheet stays present when
backgrounding the app. When the app re-enters the foreground, the
keyboard overlays the password suggestion sheet. This is a bit annoying
for users (see bug here:
https://bugs.chromium.org/p/chromium/issues/detail?id=1298500).

This CL closes the password suggestion sheet when the app enters the
background; when the app enters the foreground, only the keyboard will
be presented.

Bug: 1298500
Change-Id: I5c8ac818d5180f06f3733283428d95f56166f5bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3494879
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Auto-Submit: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#976232}
NOKEYCHECK=True
GitOrigin-RevId: 8d38bc185fc05156772b2114501ec7662d06ff68