Fix toptoolbar trailing button not responding to taps
There has been a bug in TabGridBottomToolbar for a while where
the UIView's pointInside method returns YES even if the toolbar or the
big new tab button is not on the screen.

This is a quick fix to work around that problem by making sure that
pointInside works correctly in search mode and remote tabs page.
A follow up CL will be sent to update the pointInside and whole toolbar
further and make sure that it only depends on the new tab button if and
only if this button is on the screen.

Bug: 1299421, 1298895
Change-Id: I19ae058f44f04f1e62215c3f40f089ba091b6a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3498661
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Auto-Submit: Mohammad Refaat <mrefaat@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#976428}
NOKEYCHECK=True
GitOrigin-RevId: 2d3b7da014839dffc8337a8cc23de0c91da8d21b