[ios, ksingleNTP] Return early in newTabPageHelperDidChangeVisibility:
In the instance that a pageload starts while the WebState is not the
active WebState anymore, returning early instead of setting the
NTP coordinator's WebState to nullptr will prevent a state where the NTP
is shown but it's classes has a nullptr WebState reference.

newTabPageHelperDidChangeVisibility: used to cover a case where a window
size change when showing the TabGrid with no tabs triggers an NTP
animation. Now that it will not update the NTP coordinator's WebState,
it needs to be done in this unique WebState addition instance.

Bug: 1301784
Change-Id: Ie8909b06c7d830d8ebd421edacaef9c3b6d619cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3495400
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#976450}
NOKEYCHECK=True
GitOrigin-RevId: fc42be8a49e04d000b6c5d75419422da063b18c3