Update TabGridBottomToolbar to account for floatingButton case correctly
Right now the toolbar is dealing with the floatingButton as its the
default case. Specifically in pointInside and intrinsicContentSize
check for other conditions (mode/page) or orientation and if any of
these conditions are not met it falls back to consider that the
floatingButton is available and returns value based on that.

Given that the floatingButton is a special case it shouldn't be dealt
with as the default case. Instead it should be checked independently
and if not found fallback to the super view methods regardless of the
mode or the case (unless there is another special case for that mode).

Bug: 1302116
Change-Id: I6bb7b287603503f6461b232678893929236cfcf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499343
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#976589}
NOKEYCHECK=True
GitOrigin-RevId: 53707f5b18963d0481dbb7d9d6a7698683323fff