[IOS] Add close tabs from other windows support in search mode
This is handled by checking if the tab is from another browser in the
tabgridmediator and then directing the close call to the correct
WebStateList.

Also fix a problem when Canceling the search mode doesn't clear tabs
from other windows properly if there is no tabs in the current tab grid,
This caused the application to crash if these tabs don't exist any more.

Fixed: 1302275, 1302262
Change-Id: Ia9f7ea25c4f06c1642a5a0c92ef1dcd6b7778ee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499790
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977227}
NOKEYCHECK=True
GitOrigin-RevId: e2c9acd87d698304c1aa17f04dd7530f86cc3d98