Remove gogerald@chromium.org from ios/chrome/browser/ui/start_surface/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ia69ee897958f0da6647bcccf0f2e0cce01ce071e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501564
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977229}
NOKEYCHECK=True
GitOrigin-RevId: 6c47b03521e0393d9eb6b06c3b9d9ed0ee4459ff