[iOS] Implement tap to select in the SwiftUI popup
This CL enables tapping a popup row to select the associated
autocomplete suggestion and animates the background color accordingly
when it is pressed. Tapping is properly ignored if we think the user is
tapping the trailing button or the action button instead.

Bug: 1301705
Change-Id: I663b2dcd8c64e97f216d472c2d44864fe3f3c225
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3497708
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#977680}
NOKEYCHECK=True
GitOrigin-RevId: e1a0e945e082f4bdaa2482d9137256ca4d39d7f2